# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dreamers Solutions ERP APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and obtain JWT token",
                "operationId": "9507d6722550903567336641568f1d4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "500": {
                        "description": "Could not create token"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "f1fde7f4a932e393f6042d1b92e65b0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResponse"
                                        },
                                        "token": {
                                            "description": "JWT access token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get authenticated user",
                "operationId": "68da8368cc218f46c3fd874d45d6cbbf",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "401": {
                        "description": "Token expired, invalid, or absent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/permissions/{locationId}/{moduleId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user permissions by location and module",
                "operationId": "aa838322d1a8e7ef1793808cbf3fc0d5",
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "description": "Location id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "moduleId",
                        "in": "path",
                        "description": "Module id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PermissionResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "title": "LoginRequest",
                "description": "Login credentials",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "User password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "RegisterRequest",
                "description": "User registration body",
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "description": "Full name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "User password",
                        "type": "string"
                    },
                    "password_confirmation": {
                        "description": "Password confirmation",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResponse": {
                "title": "UserResponse",
                "description": "Authenticated user response",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the user",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Full name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string"
                    },
                    "user_role_id": {
                        "description": "User role id",
                        "type": "integer"
                    },
                    "active": {
                        "description": "Active status of the user",
                        "type": "integer"
                    },
                    "token": {
                        "description": "JWT access token",
                        "type": "string"
                    },
                    "locations": {
                        "description": "Locations the user can access",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "modules": {
                        "description": "Modules the user can access",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "PermissionResponse": {
                "title": "PermissionResponse",
                "description": "Permission record",
                "properties": {
                    "id": {
                        "description": "Permission id",
                        "type": "integer"
                    },
                    "permission_code": {
                        "description": "Permission code",
                        "type": "string"
                    },
                    "fxml_path": {
                        "description": "FXML path for menu permission",
                        "type": "string"
                    },
                    "display_name": {
                        "description": "Display name",
                        "type": "string"
                    },
                    "parent_id": {
                        "description": "Parent permission id",
                        "type": "integer"
                    },
                    "depth": {
                        "description": "Permission depth in hierarchy",
                        "type": "integer"
                    },
                    "permission_type": {
                        "description": "Permission type",
                        "type": "string"
                    },
                    "active": {
                        "description": "Active status",
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from the login endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}